Move JsonLineFormatter to common
